Move resource collection and types into a group
