+++ R.swift/Collector/AssetFolder.swift
@@ -1,0 +1,35 @@
+//
+//  AssetFolder.swift
+//  R.swift
+//
+//  Created by Mathijs Kadijk on 09-12-15.
+//  Copyright © 2015 Mathijs Kadijk. All rights reserved.
+//
+
+import Foundation
+
+struct AssetFolder {
+  let name: String
+  let imageAssets: [String]
+
+  init(url: NSURL, fileManager: NSFileManager) throws {
+    guard let pathExtension = url.pathExtension where AssetFolderExtensions.contains(pathExtension) else {
+      throw ResourceParsingError.UnsupportedExtension(givenExtension: url.pathExtension, supportedExtensions: AssetFolderExtensions)
+    }
+
+    name = url.filename!
+
+    // Browse asset directory recursively and list only the assets folders
+    var assets = [NSURL]()
+    let enumerator = fileManager.enumeratorAtURL(url, includingPropertiesForKeys: nil, options: .SkipsHiddenFiles, errorHandler: nil)
+    if let enumerator = enumerator {
+      for file in enumerator {
+        if let fileURL = file as? NSURL, pathExtension = fileURL.pathExtension where AssetExtensions.indexOf(pathExtension) != nil {
+          assets.append(fileURL)
+        }
+      }
+    }
+
+    imageAssets = assets.map { $0.filename! }
+  }
+}
