+++ R.swift/Collector/Nib.swift
@@ -1,0 +1,96 @@
+//
+//  Nib.swift
+//  R.swift
+//
+//  Created by Mathijs Kadijk on 09-12-15.
+//  Copyright © 2015 Mathijs Kadijk. All rights reserved.
+//
+
+import Foundation
+
+struct Nib: ReusableContainer {
+  let name: String
+  let rootViews: [Type]
+  let reusables: [Reusable]
+
+  init(url: NSURL) throws {
+    guard let pathExtension = url.pathExtension where NibExtensions.contains(pathExtension) else {
+      throw ResourceParsingError.UnsupportedExtension(givenExtension: url.pathExtension, supportedExtensions: NibExtensions)
+    }
+
+    name = url.filename!
+
+    let parserDelegate = NibParserDelegate();
+
+    let parser = NSXMLParser(contentsOfURL: url)!
+    parser.delegate = parserDelegate
+    parser.parse()
+
+    rootViews = parserDelegate.rootViews
+    reusables = parserDelegate.reusables
+  }
+}
+
+private class NibParserDelegate: NSObject, NSXMLParserDelegate {
+  let ignoredRootViewElements = ["placeholder"]
+  var rootViews: [Type] = []
+  var reusables: [Reusable] = []
+
+  // State
+  var isObjectsTagOpened = false;
+  var levelSinceObjectsTagOpened = 0;
+
+  @objc func parser(parser: NSXMLParser, didStartElement elementName: String, namespaceURI: String?, qualifiedName qName: String?, attributes attributeDict: [String : String]) {
+    switch elementName {
+    case "objects":
+      isObjectsTagOpened = true;
+
+    default:
+      if isObjectsTagOpened {
+        levelSinceObjectsTagOpened++;
+
+        if let rootView = viewWithAttributes(attributeDict)
+          where levelSinceObjectsTagOpened == 1 && ignoredRootViewElements.filter({ $0 == elementName }).count == 0 {
+            rootViews.append(rootView)
+        }
+      }
+
+      if let reusable = reusableFromAttributes(attributeDict, elementName: elementName) {
+        reusables.append(reusable)
+      }
+    }
+  }
+
+  @objc func parser(parser: NSXMLParser, didEndElement elementName: String, namespaceURI: String?, qualifiedName qName: String?) {
+    switch elementName {
+    case "objects":
+      isObjectsTagOpened = false;
+
+    default:
+      if isObjectsTagOpened {
+        levelSinceObjectsTagOpened--;
+      }
+    }
+  }
+
+  func viewWithAttributes(attributeDict: [NSObject : AnyObject]) -> Type? {
+    let customModule = attributeDict["customModule"] as? String
+    let customClass = (attributeDict["customClass"] as? String) ?? "UIView"
+
+    return Type(module: customModule, name: customClass)
+  }
+
+  func reusableFromAttributes(attributeDict: [NSObject : AnyObject], elementName: String) -> Reusable? {
+    if let reuseIdentifier = attributeDict["reuseIdentifier"] as? String {
+      let customModule = attributeDict["customModule"] as? String
+      let customClass = attributeDict["customClass"] as? String
+      let customType = customClass.map { Type(module: customModule, name: $0, optional: false) }
+
+      let type = customType ?? ElementNameToTypeMapping[elementName] ?? Type._UIView
+
+      return Reusable(identifier: reuseIdentifier, type: type)
+    }
+    
+    return nil
+  }
+}
