+++ R.swift/types.swift
@@ -342,354 +342,3 @@ struct Struct: CustomStringConvertible {
     return "struct \(type)\(implementsString) {\n\(bodyString)\n}"
   }
 }
-
-/// MARK: Resource types
-
-enum ResourceParsingError: ErrorType {
-  case UnsupportedExtension(givenExtension: String?, supportedExtensions: Set<String>)
-  case ParsingFailed(String)
-}
-
-struct Xcodeproj {
-  private let projectFile: XCProjectFile
-  //let onDemandResourceTags: [String]
-
-  init(url: NSURL) throws {
-    // Parse project file
-    guard let projectFile = try? XCProjectFile(xcodeprojURL: url) else {
-      throw ResourceParsingError.ParsingFailed("Project file at '\(url)' could not be parsed, is this a valid Xcode project file ending in *.xcodeproj?")
-    }
-
-    self.projectFile = projectFile
-  }
-
-  func resourceURLsForTarget(targetName: String, pathResolver: Path -> NSURL) throws -> [NSURL] {
-    // Look for target in project file
-    let allTargets = projectFile.project.targets
-    guard let target = allTargets.filter({ $0.name == targetName }).first else {
-      let availableTargets = allTargets.map { $0.name }.joinWithSeparator(", ")
-      throw ResourceParsingError.ParsingFailed("Target '\(targetName)' not found in project file, available targets are: \(availableTargets)")
-    }
-
-    let resourcesFileRefs = target.buildPhases
-      .flatMap { $0 as? PBXResourcesBuildPhase }
-      .flatMap { $0.files }
-      .map { $0.fileRef }
-
-    let fileRefPaths = resourcesFileRefs
-      .flatMap { $0 as? PBXFileReference }
-      .map { $0.fullPath }
-
-    let variantGroupPaths = resourcesFileRefs
-      .flatMap { $0 as? PBXVariantGroup }
-      .flatMap { $0.fileRefs }
-      .map { $0.fullPath }
-
-    return (fileRefPaths + variantGroupPaths)
-      .map(pathResolver)
-  }
-}
-
-struct AssetFolder {
-  let name: String
-  let imageAssets: [String]
-
-  init(url: NSURL, fileManager: NSFileManager) throws {
-    guard let pathExtension = url.pathExtension where AssetFolderExtensions.contains(pathExtension) else {
-      throw ResourceParsingError.UnsupportedExtension(givenExtension: url.pathExtension, supportedExtensions: AssetFolderExtensions)
-    }
-
-    name = url.filename!
-
-    // Browse asset directory recursively and list only the assets folders
-    var assets = [NSURL]()
-    let enumerator = fileManager.enumeratorAtURL(url, includingPropertiesForKeys: nil, options: .SkipsHiddenFiles, errorHandler: nil)
-    if let enumerator = enumerator {
-      for file in enumerator {
-        if let fileURL = file as? NSURL, pathExtension = fileURL.pathExtension where AssetExtensions.indexOf(pathExtension) != nil {
-          assets.append(fileURL)
-        }
-      }
-    }
-    
-    imageAssets = assets.map { $0.filename! }
-  }
-}
-
-struct Image {
-  let name: String
-
-  init(url: NSURL) throws {
-    guard let pathExtension = url.pathExtension?.lowercaseString where ImageExtensions.contains(pathExtension) else {
-      throw ResourceParsingError.UnsupportedExtension(givenExtension: url.pathExtension, supportedExtensions: ImageExtensions)
-    }
-
-    guard let filename = url.lastPathComponent else {
-      throw ResourceParsingError.ParsingFailed("Filename could not be parsed from URL: \(url.absoluteString)")
-    }
-
-    let extensions = ImageExtensions.joinWithSeparator("|")
-    let regex = try! NSRegularExpression(pattern: "(~(ipad|iphone))?(@[2,3]x)?\\.(\(extensions))$", options: .CaseInsensitive)
-    let fullFileNameRange = NSRange(location: 0, length: filename.characters.count)
-    let pathExtensionToUse = (pathExtension == "png") ? "" : ".\(pathExtension)"
-    name = regex.stringByReplacingMatchesInString(filename, options: NSMatchingOptions(rawValue: 0), range: fullFileNameRange, withTemplate: pathExtensionToUse)
-  }
-}
-
-struct Font {
-  let name: String
-
-  init(url: NSURL) throws {
-    guard let pathExtension = url.pathExtension where FontExtensions.contains(pathExtension) else {
-      throw ResourceParsingError.UnsupportedExtension(givenExtension: url.pathExtension, supportedExtensions: FontExtensions)
-    }
-
-    let dataProvider = CGDataProviderCreateWithURL(url)
-    let font = CGFontCreateWithDataProvider(dataProvider)
-
-    guard let postScriptName = CGFontCopyPostScriptName(font) else {
-      throw ResourceParsingError.ParsingFailed("No postcriptName associated to font at \(url)")
-    }
-
-    name = postScriptName as String
-  }
-}
-
-struct Storyboard: ReusableContainer {
-  let name: String
-  private let initialViewControllerIdentifier: String?
-  let viewControllers: [ViewController]
-  let usedImageIdentifiers: [String]
-  let reusables: [Reusable]
-
-  var initialViewController: ViewController? {
-    return viewControllers.filter { $0.id == self.initialViewControllerIdentifier }.first
-  }
-
-  init(url: NSURL) throws {
-    guard let pathExtension = url.pathExtension where StoryboardExtensions.contains(pathExtension) else {
-      throw ResourceParsingError.UnsupportedExtension(givenExtension: url.pathExtension, supportedExtensions: StoryboardExtensions)
-    }
-
-    name = url.filename!
-
-    let parserDelegate = StoryboardParserDelegate()
-
-    let parser = NSXMLParser(contentsOfURL: url)!
-    parser.delegate = parserDelegate
-    parser.parse()
-
-    initialViewControllerIdentifier = parserDelegate.initialViewControllerIdentifier
-    viewControllers = parserDelegate.viewControllers
-    usedImageIdentifiers = parserDelegate.usedImageIdentifiers
-    reusables = parserDelegate.reusables
-  }
-
-  struct ViewController {
-    let id: String
-    let storyboardIdentifier: String?
-    let type: Type
-    private(set) var segues: [Segue]
-
-    mutating func addSegue(segue: Segue) {
-      segues.append(segue)
-    }
-  }
-
-  struct Segue {
-    let identifier: String
-    let type: Type
-    let destination: String
-  }
-}
-
-struct Nib: ReusableContainer {
-  let name: String
-  let rootViews: [Type]
-  let reusables: [Reusable]
-
-  init(url: NSURL) throws {
-    guard let pathExtension = url.pathExtension where NibExtensions.contains(pathExtension) else {
-      throw ResourceParsingError.UnsupportedExtension(givenExtension: url.pathExtension, supportedExtensions: NibExtensions)
-    }
-
-    name = url.filename!
-
-    let parserDelegate = NibParserDelegate();
-
-    let parser = NSXMLParser(contentsOfURL: url)!
-    parser.delegate = parserDelegate
-    parser.parse()
-
-    rootViews = parserDelegate.rootViews
-    reusables = parserDelegate.reusables
-  }
-}
-
-struct ResourceFile {
-  let fullname: String
-  let filename: String
-  let pathExtension: String?
-
-  init(url: NSURL) throws {
-    if let pathExtension = url.pathExtension where CompiledResourcesExtensions.contains(pathExtension) {
-        throw ResourceParsingError.UnsupportedExtension(givenExtension: pathExtension, supportedExtensions: ["*"])
-    }
-
-    guard let fullname = url.lastPathComponent, filename = url.filename else {
-      throw ResourceParsingError.ParsingFailed("Couldn't extract filename without extension from URL: \(url)")
-    }
-
-    self.fullname = fullname
-    self.filename = filename
-    pathExtension = url.pathExtension
-  }
-}
-
-/// MARK: Parsers
-
-class StoryboardParserDelegate: NSObject, NSXMLParserDelegate {
-  var initialViewControllerIdentifier: String?
-  var viewControllers: [Storyboard.ViewController] = []
-  var usedImageIdentifiers: [String] = []
-  var reusables: [Reusable] = []
-
-  // State
-  var currentViewController: (String, Storyboard.ViewController)?
-
-  func parser(parser: NSXMLParser, didStartElement elementName: String, namespaceURI: String?, qualifiedName qName: String?, attributes attributeDict: [String : String]) {
-    switch elementName {
-    case "document":
-      if let initialViewController = attributeDict["initialViewController"] {
-        initialViewControllerIdentifier = initialViewController
-      }
-
-    case "segue":
-      if let segueIdentifier = attributeDict["identifier"], segueDestination = attributeDict["destination"] {
-        let customModule = attributeDict["customModule"]
-        let customClass = attributeDict["customClass"]
-        let customType = customClass.map { Type(module: customModule, name: $0, optional: false) }
-
-        let type = customType ?? Type._UIStoryboardSegue
-
-        let segue = Storyboard.Segue(identifier: segueIdentifier, type: type, destination: segueDestination)
-        currentViewController!.1.addSegue(segue)
-      }
-
-    case "image":
-      if let imageIdentifier = attributeDict["name"] {
-        usedImageIdentifiers.append(imageIdentifier)
-      }
-
-    default:
-      if let viewController = viewControllerFromAttributes(attributeDict, elementName: elementName) {
-        currentViewController = (elementName, viewController)
-      }
-
-      if let reusable = reusableFromAttributes(attributeDict, elementName: elementName) {
-        reusables.append(reusable)
-      }
-    }
-  }
-
-  func parser(parser: NSXMLParser, didEndElement elementName: String, namespaceURI: String?, qualifiedName qName: String?) {
-    if let currentViewController = currentViewController where elementName == currentViewController.0 {
-      viewControllers.append(currentViewController.1)
-      self.currentViewController = nil
-    }
-  }
-
-  func viewControllerFromAttributes(attributeDict: [NSObject : AnyObject], elementName: String) -> Storyboard.ViewController? {
-    guard let id = attributeDict["id"] as? String where attributeDict["sceneMemberID"] as? String == "viewController" else {
-      return nil
-    }
-
-    let storyboardIdentifier = attributeDict["storyboardIdentifier"] as? String
-
-    let customModule = attributeDict["customModule"] as? String
-    let customClass = attributeDict["customClass"] as? String
-    let customType = customClass.map { Type(module: customModule, name: $0, optional: false) }
-
-    let type = customType ?? ElementNameToTypeMapping[elementName] ?? Type._UIViewController
-
-    return Storyboard.ViewController(id: id, storyboardIdentifier: storyboardIdentifier, type: type, segues: [])
-  }
-
-  func reusableFromAttributes(attributeDict: [NSObject : AnyObject], elementName: String) -> Reusable? {
-    guard let reuseIdentifier = attributeDict["reuseIdentifier"] as? String else {
-      return nil
-    }
-
-    let customModule = attributeDict["customModule"] as? String
-    let customClass = attributeDict["customClass"] as? String
-    let customType = customClass.map { Type(module: customModule, name: $0, optional: false) }
-
-    let type = customType ?? ElementNameToTypeMapping[elementName] ?? Type._UIView
-
-    return Reusable(identifier: reuseIdentifier, type: type)
-  }
-}
-
-class NibParserDelegate: NSObject, NSXMLParserDelegate {
-  let ignoredRootViewElements = ["placeholder"]
-  var rootViews: [Type] = []
-  var reusables: [Reusable] = []
-
-  // State
-  var isObjectsTagOpened = false;
-  var levelSinceObjectsTagOpened = 0;
-
-  func parser(parser: NSXMLParser, didStartElement elementName: String, namespaceURI: String?, qualifiedName qName: String?, attributes attributeDict: [String : String]) {
-    switch elementName {
-    case "objects":
-      isObjectsTagOpened = true;
-
-    default:
-      if isObjectsTagOpened {
-        levelSinceObjectsTagOpened++;
-
-        if let rootView = viewWithAttributes(attributeDict)
-          where levelSinceObjectsTagOpened == 1 && ignoredRootViewElements.filter({ $0 == elementName }).count == 0 {
-            rootViews.append(rootView)
-        }
-      }
-
-      if let reusable = reusableFromAttributes(attributeDict, elementName: elementName) {
-        reusables.append(reusable)
-      }
-    }
-  }
-
-  func parser(parser: NSXMLParser, didEndElement elementName: String, namespaceURI: String?, qualifiedName qName: String?) {
-    switch elementName {
-    case "objects":
-      isObjectsTagOpened = false;
-
-    default:
-      if isObjectsTagOpened {
-        levelSinceObjectsTagOpened--;
-      }
-    }
-  }
-
-  func viewWithAttributes(attributeDict: [NSObject : AnyObject]) -> Type? {
-    let customModule = attributeDict["customModule"] as? String
-    let customClass = (attributeDict["customClass"] as? String) ?? "UIView"
-    
-    return Type(module: customModule, name: customClass)
-  }
-
-  func reusableFromAttributes(attributeDict: [NSObject : AnyObject], elementName: String) -> Reusable? {
-    if let reuseIdentifier = attributeDict["reuseIdentifier"] as? String {
-      let customModule = attributeDict["customModule"] as? String
-      let customClass = attributeDict["customClass"] as? String
-      let customType = customClass.map { Type(module: customModule, name: $0, optional: false) }
-
-      let type = customType ?? ElementNameToTypeMapping[elementName] ?? Type._UIView
-
-      return Reusable(identifier: reuseIdentifier, type: type)
-    }
-
-    return nil
-  }
-}
